+++ CHANGELOG.md
@@ -9,9 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - `reserved-memory` child node types.
-- Require of `#address-cells` to be defined when using `interrupt` property.
+- Require of `#address-cells` to be defined on node parent when using `interrupt` property.
 - Require of `#address-cells` to be defined when using `interrupt-map` property.
-- Require of `#address-cells` to be defined when using `interrupts-extended` property.
+- Require of `#address-cells` to be defined on node parent when using `interrupts-extended` property.
 - Validation of `interrupt` property is linking to a `interrupt-map` entry.
 - Validation of `interrupts-extended` property is linking to a `interrupt-map` entry.
 - Validation of `interrupt-map-mask` has the correct number of `U32` values.